pesos/requirements.txt
# https://github.com/bear/python-twitter
# https://chat.indieweb.org/2016-09-16
python-twitter
easywebdav
